+++ tests/test_checksum.py
@@ -19,10 +19,14 @@
 # THE SOFTWARE.
 
 from __future__ import absolute_import
+import mock
 
 import pytest
 
 from tchannel import messages
+from tchannel import TChannel
+from tchannel import thrift
+from tchannel.errors import FatalProtocolError
 from tchannel.io import BytesIO
 from tchannel.messages import CallRequestMessage
 from tchannel.messages import ChecksumType
@@ -30,13 +34,14 @@ from tchannel.messages.common import generate_checksum
 from tchannel.messages.common import verify_checksum
 
 
-@pytest.mark.parametrize('checksum_type, seed', [
-    (ChecksumType.none, 0),
-    (ChecksumType.crc32, 0x0812fa3f),
+@pytest.mark.parametrize('checksum_type', [
+    (ChecksumType.none),
+    (ChecksumType.crc32),
+    (ChecksumType.crc32c),
 ])
-def test_checksum(checksum_type, seed):
+def test_checksum(checksum_type):
     message = CallRequestMessage()
-    message.checksum = (checksum_type, seed)
+    message.checksum = (checksum_type, None)
     generate_checksum(message)
     payload = messages.RW[message.message_type].write(
         message, BytesIO()
@@ -44,3 +49,24 @@ def test_checksum(checksum_type, seed):
 
     msg = messages.RW[message.message_type].read(BytesIO(payload))
     assert verify_checksum(msg)
+
+
+@pytest.mark.gen_test
+def test_default_checksum_type():
+    server = TChannel("server")
+    server.listen()
+    with mock.patch(
+        'tchannel.messages.common.compute_checksum', autospec=True,
+    ) as mock_compute_checksum:
+        client = TChannel("client")
+        service = thrift.load(
+            path='tchannel/health/meta.thrift',
+            service='health_test_server',
+            hostport=server.hostport,
+        )
+        with pytest.raises(FatalProtocolError):
+            yield client.thrift(service.Meta.health())
+
+        mock_compute_checksum.assert_called_with(
+            ChecksumType.crc32c, mock.ANY, mock.ANY,
+        )